Fetch access_token and render background image from artsy API
Makes 2 fetch requests, 1 to get the access token and stores that in the browser localStorage for next time and 1 request to get some actual random art with the /artworks?&sample=1 resource which grabs a random element in the JSON reponse. Right now, this component renders a backup image from Imgur when there is no artsy access token. That needs to be improved.